Serializer converts size_t to uint32_t to handle various architectures

+++ src/core/serialize/serialize_uuid.h
@@ -38,9 +38,9 @@ namespace detail
 template <class T>
 struct getSizeHelper<T, typename std::enable_if<std::is_same<T, UUID>::value>::type>
 {
-    static size_t value(const T& obj)
+    static uint32_t value(const T& obj)
     {
-        constexpr size_t size = sizeof(obj);
+        constexpr uint32_t size = sizeof(obj);
         return size;
     };
 };
@@ -50,7 +50,7 @@ struct serializeHelper<T, typename std::enable_if<std::is_same<T, UUID>::value>:
 {
     static void apply(const T& obj, std::vector<uint8_t>::iterator& it)
     {
-        constexpr size_t size = sizeof(T);
+        constexpr uint32_t size = sizeof(T);
         const uint8_t* ptr = reinterpret_cast<const uint8_t*>(&obj);
         std::copy(ptr, ptr + size, it);
         it += size;
